+++ blockindex/contractstaking/bucket_type.go
@@ -11,7 +11,7 @@ import (
 	"github.com/pkg/errors"
 	"google.golang.org/protobuf/proto"
 
-	"github.com/iotexproject/iotex-core/action/protocol/staking/contractstaking/contractstakingpb"
+	"github.com/iotexproject/iotex-core/blockindex/contractstaking/contractstakingpb"
 	"github.com/iotexproject/iotex-core/pkg/util/byteutil"
 )
