+++ blockindex/contractstaking/cache.go
