+++ blockindex/contractstaking/contractstakingpb/contractstaking.proto
@@ -7,7 +7,7 @@
 //      protoc --go_out=plugins=grpc:. *.proto
 syntax = "proto3";
 package contractstakingpb;
-option go_package = "github.com/iotexproject/iotex-core/action/protocol/staking/contractstaking/contractstakingpb";
+option go_package = "github.com/iotexproject/iotex-core/blockindex/contractstaking/contractstakingpb";
 
 message BucketType {
     string amount = 1;
