+++ blockindex/contractstaking/delta_cache.go
