+++ blockindex/contractstaking/delta_state.go
