+++ blockindex/contractstaking/indexer_test.go
