+++ blockindex/contractstaking/util.go
